Reduce allocations when filtering shape key list
Multiple allocations were made while filtering the shape key list before
the list was ever enumerated for displaying. This caused a severe drop
in frame rate when viewing this list. With these allocations removed,
the frame rate drop is not as severe and IMGUI is now the limiting
factor for performance.